Cloud-Native E-Commerce Platform/services/payment-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pika==1.3.2
kafka-python==2.0.2
prometheus-client==0.20.0
